requirements.txt
streamlit
pandas
python-dotenv
supabase